chore: unified status display system
